refactor, runnable server, configable secret
